# Use unsafe for bytes->string conversion.