Benchmark

```
benchmark                       old ns/op     new ns/op     delta
BenchmarkRedisSetGet64Bytes     103732        100645        -2.98%
BenchmarkRedisSetGet1KB         104926        101539        -3.23%
BenchmarkRedisSetGet10KB        107164        102861        -4.02%
BenchmarkRedisSetGet1MB         5206102       3871532       -25.63%

benchmark                       old allocs     new allocs     delta
BenchmarkRedisSetGet64Bytes     16             15             -6.25%
BenchmarkRedisSetGet1KB         16             15             -6.25%
BenchmarkRedisSetGet10KB        16             15             -6.25%
BenchmarkRedisSetGet1MB         17             16             -5.88%

benchmark                       old bytes     new bytes     delta
BenchmarkRedisSetGet64Bytes     720           656           -8.89%
BenchmarkRedisSetGet1KB         3664          2640          -27.95%
BenchmarkRedisSetGet10KB        3664          2640          -27.95%
BenchmarkRedisSetGet1MB         4211153       3162576       -24.90%
```

@dim?