@dim 2015-07-20
Two things I am not sure about. One is GC: if you create an unsafe reference to a byte slice and that byte slice is garbage collected, what then? Just a thought, may not be relevant. 

Also, would like to see a GET/SET benchmark with a relatively large value, eg a few MBs.