@vmihailenco 2015-07-20
I believe GC correctly handles unsafe.Pointer.

Will add more benchmarks.