@vmihailenco 2015-07-20
Updated benchmarks. See first message.